+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inria_Sans } from "next/font/google";
 import "./globals.css";
-import Providers from "./providers";
+import Providers from "../providers";
 import { cn } from "@/lib/utils";
 
 const inriaSans = Inria_Sans({
